Fixing twisted tests.
It has been observed that when a "twisted" test has "very little"
commands in the stack, it uses to fail since the proxied service
cannot be reached.

This commit puts some delay before the first command gets run
by the test interpreter; both in the form of a "blind" sleep, and
with a more structured 'wget' loop.